+++ package.json
@@ -11,7 +11,7 @@
     "lint:fix": "eslint . --fix",
     "copydev": "babel-node ./server/models/index.js createTables",
     "deletedev": "babel-node ./server/models/index.js tearDown",
-    "test": "npm run deletedev && npm run copydev && nyc --reporter=html --reporter=text mocha --require @babel/register --timeout 100000 server/tests/ --exit",
+    "test": "npm run deletedev && npm run copydev && nyc --reporter=html --reporter=text mocha --require @babel/register --timeout 100000 server/tests/main.test.js --exit",
     "production": "nodemon --exec babel-node server/index.js",
     "heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install && npm run build"
   },
