+++ server/tests/main.test.js
@@ -1,0 +1,8 @@
+import runUserTests from './user.spec';
+import runArticlesTests from './articles.spec';
+import runFlagsTests from './flags.spec';
+
+
+runUserTests();
+runArticlesTests();
+runFlagsTests();
